+++ src/main/java/io/github/seggan/sfcalc/CalcExecutor.java
@@ -1,10 +1,8 @@
 package io.github.seggan.sfcalc;
 
-import io.github.thebusybiscuit.slimefun4.core.categories.FlexCategory;
-import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
-import io.github.thebusybiscuit.slimefun4.utils.PatternUtils;
-import me.mrCookieSlime.Slimefun.Objects.Category;
-import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
+import java.util.HashMap;
+import java.util.Map;
+
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
 import org.bukkit.command.Command;
@@ -13,51 +11,55 @@ import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.Inventory;
 
-import java.util.HashMap;
-import java.util.Map;
+import io.github.thebusybiscuit.slimefun4.core.categories.FlexCategory;
+import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
+import io.github.thebusybiscuit.slimefun4.utils.PatternUtils;
+import me.mrCookieSlime.Slimefun.Objects.Category;
+import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 
+/*
+ * Copyright (C) 2020 Seggan
+ * Email: segganew@gmail.com
+ * 
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ * 
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ * 
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see <https://www.gnu.org/licenses/>.
+ */
 public class CalcExecutor implements CommandExecutor {
-//    Copyright (C) 2020 Seggan
-//    Email: segganew@gmail.com
-//
-//    This program is free software: you can redistribute it and/or modify
-//    it under the terms of the GNU General Public License as published by
-//    the Free Software Foundation, either version 3 of the License, or
-//            (at your option) any later version.
-//
-//    This program is distributed in the hope that it will be useful,
-//    but WITHOUT ANY WARRANTY; without even the implied warranty of
-//    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//    GNU General Public License for more details.
-//
-//    You should have received a copy of the GNU General Public License
-//    along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
     private final SFCalc plugin;
-
     private final Map<String, SlimefunItem[]> exceptions = new HashMap<>();
 
     public CalcExecutor(SFCalc plugin) {
         this.plugin = plugin;
 
-//        SlimefunItem[] steelPlateRecipe = new SlimefunItem[8];
-//        SlimefunItem steel = SlimefunItem.getByID("STEEL_INGOT");
-//        for (int n = 0; n < 8; n++) {
-//            steelPlateRecipe[n] = steel;
-//        }
-//        exceptions.put("steel_plate", steelPlateRecipe);
-//
-//        SlimefunItem[] reinforcedPlateRecipe = new SlimefunItem[8];
-//        SlimefunItem alloy = SlimefunItem.getByID("REINFORCED_ALLOY_INGOT");
-//        for (int n = 0; n < 8; n++) {
-//            steelPlateRecipe[n] = steel;
-//        }
-//        exceptions.put("steel_plate", steelPlateRecipe);
+        // SlimefunItem[] steelPlateRecipe = new SlimefunItem[8];
+        // SlimefunItem steel = SlimefunItem.getByID("STEEL_INGOT");
+        // for (int n = 0; n < 8; n++) {
+        // steelPlateRecipe[n] = steel;
+        // }
+        // exceptions.put("steel_plate", steelPlateRecipe);
+        //
+        // SlimefunItem[] reinforcedPlateRecipe = new SlimefunItem[8];
+        // SlimefunItem alloy = SlimefunItem.getByID("REINFORCED_ALLOY_INGOT");
+        // for (int n = 0; n < 8; n++) {
+        // steelPlateRecipe[n] = steel;
+        // }
+        // exceptions.put("steel_plate", steelPlateRecipe);
     }
 
     @Override
-    public boolean onCommand(CommandSender sender, Command command, String s, String[] args) {
-        if (!(sender instanceof Player) && s.equals("sfneeded")) {
+    public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
+        if (!(sender instanceof Player) && label.equals("sfneeded")) {
             sender.sendMessage(plugin.notAPlayerString);
             return true;
         }
@@ -65,6 +67,7 @@ public class CalcExecutor implements CommandExecutor {
         long amount;
         String reqItem;
         SlimefunItem item;
+
         if (args.length > 2) {
             return false;
         }
@@ -82,12 +85,10 @@ public class CalcExecutor implements CommandExecutor {
 
         if (args.length == 1) {
             amount = 1;
+        } else if (!PatternUtils.NUMERIC.matcher(args[1]).matches()) {
+            sender.sendMessage(plugin.noNumberString);
+            return true;
         } else {
-            if (!PatternUtils.NUMERIC.matcher(args[1]).matches()) {
-                sender.sendMessage(plugin.noNumberString);
-                return true;
-            }
-
             try {
                 amount = Long.parseLong(args[1]);
                 if (amount == 0 || amount > Integer.MAX_VALUE) {
@@ -101,7 +102,6 @@ public class CalcExecutor implements CommandExecutor {
         }
 
         reqItem = reqItem.toUpperCase();
-
         item = SlimefunItem.getByID(reqItem);
 
         if (item == null) {
@@ -109,32 +109,33 @@ public class CalcExecutor implements CommandExecutor {
             return true;
         }
 
-        SFCalc.itemsSearched.add(Util.capitalize(ChatColor.stripColor(item.getItemName())));
+        plugin.itemsSearched.add(Util.capitalize(ChatColor.stripColor(item.getItemName())));
 
-        Calculator.printResults(Calculator.calculate(item, plugin), sender, s, item, amount, plugin);
+        Calculator calculator = new Calculator(plugin);
+        calculator.printResults(sender, label, item, amount);
 
         return true;
     }
 
     private void openGUI(Player player) {
         int size = Util.getSlots(SlimefunPlugin.getRegistry().getCategories().size());
+
         if (size > 54) {
             player.sendMessage(plugin.tooManyCategoriesString);
             return;
         }
-        Inventory inv = Bukkit.createInventory(
-                null,
-                size,
-                "Choose a Category"
-        );
+
+        Inventory inv = Bukkit.createInventory(null, size, "Choose a Category");
+
         for (Category category : SlimefunPlugin.getRegistry().getCategories()) {
             if (category instanceof FlexCategory) {
                 continue;
             }
+
             inv.addItem(category.getItem(player));
         }
+
         player.openInventory(inv);
     }
-
 
 }