+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## Build 13:
+- Many code reformats
+- Speed improvements
+- Memory usage has been cut by a ton
+- Bugfixes
+- The license is now included in the jar
+- Entries are now listed from lowest to highest
+
 ## Build 12:
 - The calculator now shows correct amounts of items wen large amounts are required
 - Added (broken?) support for LiteXpansion metal forge